--- packages/cli/src/cmds/deploy.ts
@@ -1,60 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-import { FxError, Result, err, ok } from "@microsoft/teamsfx-api";
-import path from "path";
-import { Argv } from "yargs";
-import activate from "../activate";
-import { EnvOptions, RootFolderOptions } from "../constants";
-import { strings } from "../resource";
-import CliTelemetry, { makeEnvRelatedProperty } from "../telemetry/cliTelemetry";
-import {
-  TelemetryEvent,
-  TelemetryProperty,
-  TelemetrySuccess,
-} from "../telemetry/cliTelemetryEvents";
-import { getSystemInputs } from "../utils";
-import { YargsCommand } from "../yargsCommand";
-
-export default class Deploy extends YargsCommand {
-  public readonly commandHead = `deploy`;
-  public readonly command = this.commandHead;
-  public readonly description = strings.command.deploy.description;
-
-  public builder(yargs: Argv): Argv<any> {
-    return yargs.hide("interactive").version(false).options(EnvOptions).options(RootFolderOptions);
-  }
-
-  public async runCommand(args: { [argName: string]: string }): Promise<Result<null, FxError>> {
-    const rootFolder = path.resolve((args.folder as string) || "./");
-    CliTelemetry.withRootFolder(rootFolder).sendTelemetryEvent(TelemetryEvent.DeployStart);
-
-    const result = await activate(rootFolder);
-    if (result.isErr()) {
-      CliTelemetry.sendTelemetryErrorEvent(TelemetryEvent.Deploy, result.error);
-      return err(result.error);
-    }
-
-    const core = result.value;
-
-    const inputs = getSystemInputs(rootFolder, args.env);
-    {
-      const result = await core.deployArtifacts(inputs);
-      if (result.isErr()) {
-        CliTelemetry.sendTelemetryErrorEvent(
-          TelemetryEvent.Deploy,
-          result.error,
-          makeEnvRelatedProperty(rootFolder, inputs)
-        );
-
-        return err(result.error);
-      }
-    }
-
-    CliTelemetry.sendTelemetryEvent(TelemetryEvent.Deploy, {
-      [TelemetryProperty.Success]: TelemetrySuccess.Yes,
-      ...makeEnvRelatedProperty(rootFolder, inputs),
-    });
-    return ok(null);
-  }
-}
