--- packages/cli/src/cmds/package.ts
@@ -1,65 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-import { FxError, Result, err, ok } from "@microsoft/teamsfx-api";
-import path from "path";
-import { Argv } from "yargs";
-import activate from "../activate";
-import { BuildPackageOptions, EnvOptions, RootFolderOptions } from "../constants";
-import CliTelemetry, { makeEnvRelatedProperty } from "../telemetry/cliTelemetry";
-import {
-  TelemetryEvent,
-  TelemetryProperty,
-  TelemetrySuccess,
-} from "../telemetry/cliTelemetryEvents";
-import { getSystemInputs } from "../utils";
-import { YargsCommand } from "../yargsCommand";
-import { globals } from "../globals";
-
-export default class Package extends YargsCommand {
-  public readonly commandHead = `package`;
-  public readonly command = `${this.commandHead}`;
-  public readonly description = "Build your Teams app into a package for publishing.";
-
-  public builder(yargs: Argv): Argv<any> {
-    globals.options = ["manifest-path", "env"];
-    return yargs
-      .hide("interactive")
-      .version(false)
-      .options(RootFolderOptions)
-      .options(BuildPackageOptions)
-      .options(EnvOptions);
-  }
-
-  public async runCommand(args: { [argName: string]: string }): Promise<Result<null, FxError>> {
-    const rootFolder = path.resolve((args.folder as string) || "./");
-    CliTelemetry.withRootFolder(rootFolder).sendTelemetryEvent(TelemetryEvent.BuildStart);
-
-    const result = await activate(rootFolder);
-    if (result.isErr()) {
-      CliTelemetry.sendTelemetryErrorEvent(TelemetryEvent.Build, result.error);
-      return err(result.error);
-    }
-    const core = result.value;
-    const inputs = getSystemInputs(rootFolder, args.env);
-    inputs.ignoreEnvInfo = false;
-    {
-      const result = await core.createAppPackage(inputs);
-      if (result.isErr()) {
-        CliTelemetry.sendTelemetryErrorEvent(
-          TelemetryEvent.Build,
-          result.error,
-          makeEnvRelatedProperty(rootFolder, inputs)
-        );
-
-        return err(result.error);
-      }
-    }
-
-    CliTelemetry.sendTelemetryEvent(TelemetryEvent.Build, {
-      [TelemetryProperty.Success]: TelemetrySuccess.Yes,
-      ...makeEnvRelatedProperty(rootFolder, inputs),
-    });
-    return ok(null);
-  }
-}
