+++ packages/cli/src/cmds/preview/previewEnv.ts
@@ -8,28 +8,24 @@ import {
   AppStudioScopes,
   assembleError,
   CoreQuestionNames,
-  environmentManager,
+  environmentNameManager,
   envUtil,
   FxCore,
   getSideloadingStatus,
   HubTypes,
   loadTeamsFxDevScript,
-  TelemetryContext,
 } from "@microsoft/teamsfx-core";
 import fs from "fs-extra";
 import * as path from "path";
 import * as util from "util";
-import { Argv } from "yargs";
 import activate from "../../activate";
 import { signedOut } from "../../commonlib/common/constant";
 import cliLogger from "../../commonlib/log";
 import M365TokenInstance from "../../commonlib/m365Login";
-import { cliSource, RootFolderOptions } from "../../constants";
-import cliTelemetry from "../../telemetry/cliTelemetry";
-import { TelemetryEvent, TelemetryProperty } from "../../telemetry/cliTelemetryEvents";
+import { cliSource } from "../../constants";
+import { TelemetryEvent } from "../../telemetry/cliTelemetryEvents";
 import CLIUIInstance from "../../userInteraction";
-import { getColorizedString, getSystemInputs, isWorkspaceSupported } from "../../utils";
-import { YargsCommand } from "../../yargsCommand";
+import { getColorizedString, getSystemInputs } from "../../utils";
 import * as commonUtils from "./commonUtils";
 import * as constants from "./constants";
 import * as errors from "./errors";
@@ -37,7 +33,6 @@ import { openHubWebClientNew } from "./launch";
 import { localTelemetryReporter } from "./localTelemetryReporter";
 import { ServiceLogWriter } from "./serviceLogWriter";
 import { Task } from "./task";
-import { environmentNameManager } from "@microsoft/teamsfx-core";
 enum Progress {
   M365Account = "Microsoft 365 Account",
 }
@@ -47,122 +42,9 @@ const ProgressMessage: { [key: string]: string } = Object.freeze({
 });
 
 // The new preview cmd `teamsfx preview --env ...`
-export default class PreviewEnv extends YargsCommand {
-  public readonly commandHead = `preview`;
-  public readonly command = `${this.commandHead}`;
-  public readonly description = "Preview the current application.";
-
+export default class PreviewEnv {
   protected runningTasks: Task[] = [];
-
   private readonly telemetryProperties: { [key: string]: string } = {};
-  private readonly telemetryMeasurements: { [key: string]: number } = {};
-
-  public builder(yargs: Argv): Argv<any> {
-    yargs
-      .options(RootFolderOptions)
-      .options("env", {
-        description: "Select an existing env for the project",
-        string: true,
-        default: environmentNameManager.getLocalEnvName(),
-      })
-      .options("manifest-file-path", {
-        description:
-          "Select the Teams app manifest file path, defaults to '${folder}/appPackage/manifest.json'",
-        string: true,
-      })
-      .options("run-command", {
-        description:
-          "The command to start local service. Work for 'local' environment only. If undefined, teamsfx will use the auto detected one from project type (`npm run dev:teamsfx` or `dotnet run` or `func start`). If empty, teamsfx will skip starting local service.",
-        string: true,
-      })
-      .options("running-pattern", {
-        description: `The ready signal output that service is launched. Work for 'local' environment only. If undefined, teamsfx will use the default common pattern ("${constants.defaultRunningPattern.source}"). If empty, teamsfx treats process start as ready signal.`,
-        string: true,
-      })
-      .options("open-only", {
-        description:
-          "Work for 'local' environment only. If true, directly open web client without launching local service.",
-        boolean: true,
-        default: false,
-      })
-      .options("m365-host", {
-        description: "Preview the application in Teams, Outlook or the Microsoft 365 app",
-        string: true,
-        choices: [HubTypes.teams, HubTypes.outlook, HubTypes.office],
-        default: HubTypes.teams,
-      })
-      .options("browser", {
-        description: "Select browser to open Teams web client",
-        string: true,
-        choices: [constants.Browser.chrome, constants.Browser.edge, constants.Browser.default],
-        default: constants.Browser.default,
-      })
-      .options("browser-arg", {
-        description: 'Argument to pass to the browser (e.g. --browser-args="--guest")',
-        string: true,
-        array: true,
-      })
-      .options("exec-path", {
-        description:
-          'The paths that will be added to the system environment variable PATH when the command is executed, defaults to "${folder}/devTools/func".',
-        string: true,
-        default: constants.defaultExecPath,
-      });
-    return yargs.version(false);
-  }
-
-  public async runCommand(args: {
-    [argName: string]: boolean | string | string[] | undefined;
-  }): Promise<Result<null, FxError>> {
-    if (args.folder === undefined || !isWorkspaceSupported(args.folder as string)) {
-      return err(errors.WorkspaceNotSupported(args.folder as string));
-    }
-    const workspaceFolder = path.resolve(args.folder as string);
-    const env = args.env as string;
-    const manifestFilePath =
-      (args["manifest-file-path"] as string) ??
-      path.join(workspaceFolder, "appPackage", "manifest.json");
-    const runCommand: string | undefined = args["run-command"] as string;
-    const runningPattern = args["running-pattern"] as string;
-    const openOnly = args["open-only"] as boolean;
-    const m365Host = args["m365-host"] as HubTypes;
-    const execPath: string = args["exec-path"] as string;
-    const browser = args.browser as constants.Browser;
-    const browserArguments = (args["browser-arg"] as string[]) ?? [];
-
-    cliTelemetry.withRootFolder(workspaceFolder);
-    this.telemetryProperties[TelemetryProperty.PreviewType] =
-      environmentNameManager.isRemoteEnvironment(env.toLowerCase())
-        ? `remote-${env}`
-        : env.toLowerCase();
-    this.telemetryProperties[TelemetryProperty.PreviewHub] = m365Host;
-    this.telemetryProperties[TelemetryProperty.PreviewBrowser] = browser;
-
-    return await localTelemetryReporter.runWithTelemetryGeneric(
-      TelemetryEvent.Preview,
-      async () =>
-        this.doPreview(
-          workspaceFolder,
-          env,
-          manifestFilePath,
-          runCommand,
-          runningPattern,
-          openOnly,
-          m365Host,
-          browser,
-          browserArguments,
-          execPath
-        ),
-      (result: Result<null, FxError>, ctx: TelemetryContext) => {
-        // whether on success or failure, send this.telemetryProperties and this.telemetryMeasurements
-        Object.assign(ctx.properties, this.telemetryProperties);
-        Object.assign(ctx.measurements, this.telemetryMeasurements);
-        return result.isErr() ? result.error : undefined;
-      },
-      this.telemetryProperties
-    );
-  }
-
   async doPreview(
     workspaceFolder: string,
     env: string,
