+++ packages/cli/src/commands/models/accountLoginM365.ts
@@ -2,8 +2,8 @@
 // Licensed under the MIT license.
 import { CLICommand, ok } from "@microsoft/teamsfx-api";
 import M365TokenProvider from "../../commonlib/m365Login";
-import { outputM365Info } from "../../cmds/account";
 import { TelemetryEvent } from "../../telemetry/cliTelemetryEvents";
+import { outputM365Info } from "./accountShow";
 
 export const accountLoginM365Command: CLICommand = {
   name: "m365",
@@ -11,7 +11,7 @@ export const accountLoginM365Command: CLICommand = {
   telemetry: {
     event: TelemetryEvent.AccountLoginM365,
   },
-  handler: async (ctx) => {
+  handler: async () => {
     await M365TokenProvider.signout();
     await outputM365Info("login");
     return ok(undefined);
