+++ packages/cli/src/commands/models/index.ts
@@ -9,9 +9,6 @@ export * from "./accountLogout";
 export * from "./accountShow";
 export * from "./add";
 export * from "./addSPFxWebpart";
-export * from "./config";
-export * from "./configGet";
-export * from "./configSet";
 export * from "./create";
 export * from "./createSample";
 export * from "./deploy";
