+++ packages/cli/tests/unit/cmds/preview/previewEnv.tests.ts
@@ -2,28 +2,27 @@
 // Licensed under the MIT license.
 
 import { err, FxError, IProgressHandler, ok, Result } from "@microsoft/teamsfx-api";
-import { FxCore, envUtil, VersionCheckRes, VersionState, HubTypes } from "@microsoft/teamsfx-core";
+import { envUtil, FxCore, HubTypes, VersionCheckRes, VersionState } from "@microsoft/teamsfx-core";
 import * as packageJson from "@microsoft/teamsfx-core/build/common/local/packageJsonHelper";
-import { Hub } from "@microsoft/teamsfx-core/build/common/m365/constants";
+import * as tools from "@microsoft/teamsfx-core/build/common/tools";
 import fs from "fs-extra";
-import * as path from "path";
 import { RestoreFn } from "mocked-env";
+import * as path from "path";
 import sinon from "sinon";
 import yargs, { Options } from "yargs";
+import * as commonUtils from "../../../../src/cmds/preview/commonUtils";
 import * as constants from "../../../../src/cmds/preview/constants";
+import * as launch from "../../../../src/cmds/preview/launch";
 import PreviewEnv from "../../../../src/cmds/preview/previewEnv";
+import { ServiceLogWriter } from "../../../../src/cmds/preview/serviceLogWriter";
+import { Task } from "../../../../src/cmds/preview/task";
+import { signedIn, signedOut } from "../../../../src/commonlib/common/constant";
 import cliLogger from "../../../../src/commonlib/log";
+import M365TokenInstance from "../../../../src/commonlib/m365Login";
 import cliTelemetry from "../../../../src/telemetry/cliTelemetry";
 import CLIUIInstance from "../../../../src/userInteraction";
 import * as Utils from "../../../../src/utils";
 import { expect } from "../../utils";
-import * as commonUtils from "../../../../src/cmds/preview/commonUtils";
-import * as launch from "../../../../src/cmds/preview/launch";
-import { ServiceLogWriter } from "../../../../src/cmds/preview/serviceLogWriter";
-import { Task } from "../../../../src/cmds/preview/task";
-import M365TokenInstance from "../../../../src/commonlib/m365Login";
-import { signedIn, signedOut } from "../../../../src/commonlib/common/constant";
-import * as tools from "@microsoft/teamsfx-core/build/common/tools";
 
 describe("Preview --env", () => {
   const sandbox = sinon.createSandbox();
@@ -85,20 +84,6 @@ describe("Preview --env", () => {
   afterEach(() => {
     sandbox.restore();
     mockedEnvRestore();
-  });
-
-  it("Builder Check", () => {
-    const cmd = new PreviewEnv();
-    cmd.builder(yargs);
-
-    expect(options).includes("folder", JSON.stringify(options));
-    expect(options).includes("env", JSON.stringify(options));
-    expect(options).includes("manifest-file-path", JSON.stringify(options));
-    expect(options).includes("run-command", JSON.stringify(options));
-    expect(options).includes("running-pattern", JSON.stringify(options));
-    expect(options).includes("m365-host", JSON.stringify(options));
-    expect(options).includes("browser", JSON.stringify(options));
-    expect(options).includes("browser-arg", JSON.stringify(options));
   });
 
   it("Preview Command Running - Default", async () => {
